main: Be more verbose if quitting early from misconfiguration
If the driver lacks support for usefully running the health monitor,
be verbose by default about why we are exiting.  The resulting logs
will be handy in pointing the driver developer what thy might have
done wrong in their CSI implementation, even if they didn't pass in a
non-default --v=2.

Fixes: https://github.com/kubernetes-csi/external-health-monitor/issues/305
Signed-off-by: Eric Blake <eblake@redhat.com>